refactor: untangled cyclic dependencies
- moved `getNamedArgs/getFormatterArgs/getStringValue` to value helper
- moved `listOptions/groupValues` to `pseudo-states` as they are only used there
